+++ Packages/net.nekobako.bone-weight-modifier/Editor/PreviewUtils.cs
@@ -1,3 +1,4 @@
+using System;
 using UnityEngine;
 using nadena.dev.ndmf.preview;
 
@@ -27,6 +28,49 @@ namespace net.nekobako.BoneWeightModifier.Editor
 
             context.GetComponent<T>(component.gameObject);
             return component.TryGetComponent(out result);
+        }
+
+        public static Mesh GetSharedMesh(this ComputeContext context, Renderer renderer)
+        {
+            if (!renderer)
+            {
+                return null;
+            }
+
+            return renderer switch
+            {
+                MeshRenderer when context.TryGetComponent<MeshFilter>(renderer, out var meshFilter) => context.Observe(meshFilter, x => x.sharedMesh),
+                SkinnedMeshRenderer skinnedMeshRenderer => context.Observe(skinnedMeshRenderer, y => y.sharedMesh),
+                _ => null,
+            };
+        }
+
+        public static Transform[] GetBones(this ComputeContext context, Renderer renderer)
+        {
+            if (!renderer)
+            {
+                return Array.Empty<Transform>();
+            }
+
+            return renderer switch
+            {
+                SkinnedMeshRenderer skinnedMeshRenderer => context.Observe(skinnedMeshRenderer, y => y.bones),
+                _ => Array.Empty<Transform>(),
+            };
+        }
+
+        public static Transform GetRootBone(this ComputeContext context, Renderer renderer)
+        {
+            if (!renderer)
+            {
+                return null;
+            }
+
+            return renderer switch
+            {
+                SkinnedMeshRenderer skinnedMeshRenderer => context.Observe(skinnedMeshRenderer, y => y.rootBone),
+                _ => null,
+            };
         }
     }
 }